backend/requirements.txt
fastapi>=0.100.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
sentry-sdk>=1.30.0
slowapi>=0.1.8
fastapi-pagination>=0.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
httpx>=0.24.0
aiofiles>=23.2.1
python-multipart>=0.0.6
gunicorn>=21.2.0
uvicorn[standard]>=0.23.0
aiokafka>=0.8.1
boto3>=1.34.0
